requirements.txt
aiogram==3.3.0
aiohttp==3.9.0
python-dotenv==1.0.0
pydantic==2.5.0
apscheduler==3.10.4
Pillow==10.1.0